requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
bcrypt==3.1.7
blinker==1.4
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Mail==0.9.1
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.1
flask-swagger==0.2.14
Flask-WTF==0.14.3
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
more-itertools==8.2.0
mysqlclient==1.4.6
packaging==20.3
passlib==1.7.2
pluggy==0.13.1
py==1.8.1
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.1
python-dotenv==0.12.0
PyYAML==5.3.1
redis==3.4.1
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.16
urllib3==1.25.9
wcwidth==0.1.9
Werkzeug==1.0.1
WTForms==2.2.1
zipp==3.1.0
